requirements.txt
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
blessed==1.20.0
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==5.1.1
Django==5.1.2
django-cron==0.6.0
django-crontab==0.7.1
django-debug-toolbar==4.4.6
django-filter==24.3
django-picklefield==3.2
django-q==1.3.9
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.1.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.28.0
jedi==0.19.1
Markdown==3.7
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pygal==3.0.5
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.3
pytest-benchmark==4.0.0
pytest-fixture-classes==1.0.3
python-dateutil==2.9.0.post0
redis==3.5.3
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
wheel==0.44.0
zipp==3.20.2
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
blessed==1.20.0
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==5.1.1
Django==5.1.2
django-cron==0.6.0
django-crontab==0.7.1
django-debug-toolbar==4.4.6
django-filter==24.3
django-picklefield==3.2
django-q==1.3.9
django-q2==1.7.3
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.1.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.28.0
jedi==0.19.1
Markdown==3.7
matplotlib-inline==0.1.7
packaging @ file:///croot/packaging_1720101850331/work
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pygal==3.0.5
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.3
pytest-benchmark==4.0.0
pytest-fixture-classes==1.0.3
python-dateutil==2.9.0.post0
redis==3.5.3
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
zipp==3.20.2
